+++ docs/changes.md
@@ -5,6 +5,7 @@
   * 完善mmap.c(感谢陈谭提供补丁)
   * 完善release脚本拷贝自定义控件的动态库。
   * 完善file browser view的value changed事件(感谢智明提供补丁) 
+  * self layout default的百分比使用浮点数。
 
 2021/06/06
   * 重构widget部分代码。